Update README.md
This is how I fixed the issue on an old database that has extensions installed in another schema. Not sure if this is really a nice/elegant way of fixing this.

Maybe I should list the used functions, or link to an external file/readme with the code or solution to fix this? I don't really want to pollute the main readme with edge cases.

What's your thought on the best place to put this info?